pyproject.toml
[project]
name = "dependency-doctor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.2.1",
    "fastapi>=0.116.1",
    "pip-audit>=2.9.0",
    "pygithub>=2.6.1",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "rio-ui>=0.11.1",
    "ruff>=0.12.5",
    "structlog>=25.4.0",
    "toml>=0.10.2",
]
